Add jwt-go to deps so we can build dcos-signal-enterprise
Since we use Go Modules now we need to add all the deps to the go.mod
file since we can't just manually copy stuff to the vendor folder
because the Go compiler won't find it if it isn't mentioned in go.mod.

This commit will be accompanied by a commit in dcos-signal-enterprise
that removed the dep from the vendor folder there.